+++ github-runner-manager/src/github_runner_manager/errors.py
@@ -29,22 +29,6 @@ class GithubMetricsError(Exception):
     """Base class for all github metrics errors."""
 
 
-class PlatformClientError(Exception):
-    """Base class for all github client errors."""
-
-
-class PlatformApiError(PlatformClientError):
-    """Represents an error when the GitHub API returns an error."""
-
-
-class TokenError(PlatformClientError):
-    """Represents an error when the token is invalid or has not enough permissions."""
-
-
-class JobNotFoundError(PlatformClientError):
-    """Represents an error when the job could not be found on the platform."""
-
-
 class CloudError(Exception):
     """Base class for cloud (as e.g. OpenStack) errors."""